+++ change_log.md
@@ -13,7 +13,12 @@ dubbogo 目前版本(0.1.1)支持的codec 是jsonrpc 2.0，transport protocol是
 ### 2018-05-07
 ---
 - 1 fmt.Errorf -> juju/errors.Errorf
+
 - 2 delete ListService
+
+- 3 Registry.GetService -> Registry.GetServices
+
+  
 
 ### 2016-10-26 ###
 ---